Refactor common functions between tests and cdcpd
- Refactors some of the functions necessary for both testing as well as
some of the functions for setting up tracking. Some functions were
moved from cdcpd_node to cdcpd/src/utils.cpp
- I imagine the name for utils.cpp will change as it evolves.
