+++ cdcpd/src/cdcpd_node.cpp
@@ -23,6 +23,7 @@
 #include <arc_utilities/ros_helpers.hpp>
 
 #include "cdcpd_ros/camera_sub.h"
+#include "cdcpd/utils.h"
 
 std::string const LOGNAME = "cdcpd_node";
 constexpr auto const MAX_CONTACTS_VIZ = 25;
@@ -40,38 +41,6 @@ typedef pcl::PointCloud<pcl::PointXYZ> PointCloud;
 namespace gm = geometry_msgs;
 namespace vm = visualization_msgs;
 namespace ehc = EigenHelpersConversions;
-
-std::pair<Eigen::Matrix3Xf, Eigen::Matrix2Xi> makeRopeTemplate(int num_points, const Eigen::Vector3f& start_position,
-                                                               const Eigen::Vector3f& end_position);
-
-std::pair<Eigen::Matrix3Xf, Eigen::Matrix2Xi> makeRopeTemplate(int const num_points,
-                                                               const Eigen::Vector3f& start_position,
-                                                               const Eigen::Vector3f& end_position) {
-  Eigen::Matrix3Xf template_vertices(3, num_points);  // Y^0 in the paper
-  Eigen::VectorXf thetas = Eigen::VectorXf::LinSpaced(num_points, 0, 1);
-  for (auto i = 0u; i < num_points; ++i) {
-    auto const theta = thetas.row(i);
-    template_vertices.col(i) = (end_position - start_position) * theta + start_position;
-  }
-  Eigen::Matrix2Xi template_edges(2, num_points - 1);
-  template_edges(0, 0) = 0;
-  template_edges(1, template_edges.cols() - 1) = num_points - 1;
-  for (int i = 1; i <= template_edges.cols() - 1; ++i) {
-    template_edges(0, i) = i;
-    template_edges(1, i - 1) = i;
-  }
-  return {template_vertices, template_edges};
-}
-
-PointCloud::Ptr makeCloud(Eigen::Matrix3Xf const& points) {
-  // TODO: Can we do this cleaner via some sort of data mapping?
-  PointCloud::Ptr cloud(new PointCloud);
-  for (int i = 0; i < points.cols(); ++i) {
-    auto const& c = points.col(i);
-    cloud->push_back(pcl::PointXYZ(c(0), c(1), c(2)));
-  }
-  return cloud;
-}
 
 cv::Mat getHsvMask(ros::NodeHandle const& ph, cv::Mat const& rgb) {
   auto const hue_min = ROSHelpers::GetParamDebugLog<double>(ph, "hue_min", 340.0);
