+++ cdcpd/src/utils.cpp
@@ -1,0 +1,32 @@
+#include "cdcpd/utils.h"
+
+std::pair<Eigen::Matrix3Xf, Eigen::Matrix2Xi> makeRopeTemplate(int const num_points,
+                                                               const Eigen::Vector3f& start_position,
+                                                               const Eigen::Vector3f& end_position)
+{
+  Eigen::Matrix3Xf template_vertices(3, num_points);  // Y^0 in the paper
+  Eigen::VectorXf thetas = Eigen::VectorXf::LinSpaced(num_points, 0, 1);
+  for (int i = 0; i < num_points; ++i) {
+    auto const theta = thetas.row(i);
+    template_vertices.col(i) = (end_position - start_position) * theta + start_position;
+  }
+  Eigen::Matrix2Xi template_edges(2, num_points - 1);
+  template_edges(0, 0) = 0;
+  template_edges(1, template_edges.cols() - 1) = num_points - 1;
+  for (int i = 1; i <= template_edges.cols() - 1; ++i) {
+    template_edges(0, i) = i;
+    template_edges(1, i - 1) = i;
+  }
+  return {template_vertices, template_edges};
+}
+
+PointCloud::Ptr makeCloud(Eigen::Matrix3Xf const& points)
+{
+  // TODO: Can we do this cleaner via some sort of data mapping?
+  PointCloud::Ptr cloud(new PointCloud);
+  for (int i = 0; i < points.cols(); ++i) {
+    auto const& c = points.col(i);
+    cloud->push_back(pcl::PointXYZ(c(0), c(1), c(2)));
+  }
+  return cloud;
+}
